requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.4
cryptography==41.0.7
databases==0.9.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-mail==1.4.1
fastapi-pagination==0.12.24
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
pytest==7.4.3
pytest-mock==3.12.0
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
schedule==1.2.1
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.32.0.post1
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0